extract pure function
