+++ src/photo/PhotoApp.jsx
@@ -126,10 +126,8 @@ class PhotoApp extends React.Component {
                   alignItems: 'center',
                   justifyContent: 'center'
                 }}
-                setPhotoInfo={this.props.setPhotoInfo}
                 media={this.props.media}
                 lookPhotoDetail={this.lookPhotoDetail}
-                getTimeline={this.props.getTimeline}
                 ipcRenderer={this.props.ipcRenderer}
                 addListToSelection={this.props.addListToSelection}
                 removeListToSelection={this.props.removeListToSelection}
