+++ src/view/Media.jsx
@@ -8,7 +8,6 @@ import NavigationMenu from 'material-ui/svg-icons/navigation/menu'
 
 import Base from './Base'
 import PhotoApp from '../photo/PhotoApp'
-import { formatDate } from '../common/datetime'
 import FlatButton from '../common/FlatButton'
 import { combineElement, removeElement } from '../common/array'
 
@@ -78,191 +77,6 @@ class Media extends Base {
 
     this.memoizeValue = { currentDigest: '', currentScrollTop: 0 }
     this.firstSelect = true
-
-    this.height = 0
-    this.width = 0
-    this.size = 218
-    this.allPhotos = []
-    this.photoDates = []
-    this.photoMapDates = []
-    this.photoListWithSameDate = []
-    this.allHeight = []
-    this.rowHeightSum = 0
-    this.indexHeightSum = []
-    this.maxScrollTop = 0
-    this.previousIndex = 1
-
-    this.setPhotoInfo = (height, width, media) => {
-      const size = Math.floor((width - 60) / Math.floor((width - 60) / 218 + 0.5))
-
-      /* mediaStore were sorted by date in Node */
-      if ((this.allPhotos !== media || this.width !== width) && width) {
-        /* init */
-        this.width = width
-        this.allPhotos = media
-        this.photoDates = []
-        this.photoMapDates = []
-        this.photoListWithSameDate = []
-        this.allHeight = []
-        this.rowHeightSum = 0
-        this.indexHeightSum = []
-        this.maxScrollTop = 0
-        this.previousIndex = 1
-
-        /* calculate photoMapDates and photoDates */
-        const MAX = Math.floor((width - 60) / size) - 1
-        let MaxItem = MAX
-        let lineIndex = 0
-        const dateUnknown = []
-        this.allPhotos.forEach((item) => {
-          const date = item.date || item.datetime
-          if (!date || date.search(/:/g) !== 4 || date.search(/^0/) > -1) { // only allow format: "2017:06:17 17:31:18"
-            dateUnknown.push(item)
-            return
-          }
-          const formatExifDateTime = formatDate(date)
-          const isRepeat = this.photoDates[this.photoDates.length - 1] === formatExifDateTime
-          if (!isRepeat || MaxItem === 0) {
-            MaxItem = MAX
-            this.photoDates.push(formatExifDateTime)
-            this.photoMapDates.push({
-              first: !isRepeat,
-              index: lineIndex,
-              date: formatExifDateTime,
-              photos: [item]
-            })
-            if (!isRepeat) {
-              this.photoListWithSameDate.push({
-                date: formatExifDateTime,
-                photos: [item]
-              })
-            } else {
-              this.photoListWithSameDate[this.photoListWithSameDate.length - 1].photos.push(item)
-            }
-            lineIndex += 1
-          } else {
-            MaxItem -= 1
-            this.photoMapDates[this.photoMapDates.length - 1].photos.push(item)
-            this.photoListWithSameDate[this.photoListWithSameDate.length - 1].photos.push(item)
-          }
-        })
-        const dateUnknownText = i18n.__('Date Unknown Text')
-        if (dateUnknown.length > 0) {
-          this.photoListWithSameDate.push({ date: dateUnknownText, photos: [] })
-          MaxItem = 0
-          lineIndex += 1
-          let isRepeat = false
-          dateUnknown.forEach((item) => {
-            this.photoListWithSameDate[this.photoListWithSameDate.length - 1].photos.push(item)
-            if (MaxItem === 0) {
-              MaxItem = MAX
-              this.photoDates.push(0)
-              this.photoMapDates.push({
-                first: !isRepeat,
-                index: lineIndex,
-                date: dateUnknownText,
-                photos: [item]
-              })
-              lineIndex += 1
-              isRepeat = true
-            } else {
-              MaxItem -= 1
-              this.photoMapDates[this.photoMapDates.length - 1]
-                .photos
-                .push(item)
-            }
-          })
-        }
-
-        /* calculate each row's heigth and their sum */
-        this.photoMapDates.forEach((list) => {
-          const tmp = size * Math.ceil(list.photos.length / Math.floor((width - 60) / size)) + !!list.first * 48
-          this.allHeight.push(tmp)
-          this.rowHeightSum += tmp
-          this.indexHeightSum.push(this.rowHeightSum)
-        })
-
-        this.maxScrollTop = this.rowHeightSum - height + 16 * 2
-        if (this.maxScrollTop > 1500000) {
-          const r = this.maxScrollTop / 1500000
-          this.indexHeightSum = this.indexHeightSum.map(h => h / r)
-          this.maxScrollTop = 1500000
-        }
-      }
-      return {
-        photoDates: this.photoDates,
-        photoMapDates: this.photoMapDates,
-        indexHeightSum: this.indexHeightSum,
-        allHeight: this.allHeight,
-        maxScrollTop: this.maxScrollTop,
-        rowHeightSum: this.rowHeightSum,
-        photoListWithSameDate: this.photoListWithSameDate,
-        size: size - 8
-      }
-    }
-
-    this.getTimeline = (photoDates, indexHeightSum, maxScrollTop, height) => {
-      const month = new Map()
-      let dateUnknown = 0
-      /* parse data to list of month */
-      photoDates.forEach((date) => {
-        if (!date) return (dateUnknown += 1)
-        const b = date.split(/-/)
-        const mix = `${b[0]}-${b[1]}`
-        if (month.has(mix)) {
-          month.set(mix, month.get(mix) + 1)
-        } else {
-          month.set(mix, 1)
-        }
-        return null
-      })
-      if (dateUnknown) month.set('0', dateUnknown)
-
-      let sumCount = 0
-      let spacingCount = 0
-      let currentYear = null
-      const timeline = [...month].map((data, index) => {
-        let percentage = 0
-        if (sumCount) {
-          percentage = (indexHeightSum[sumCount - 1] - 200) / maxScrollTop
-        }
-        /* top = percentage * height + headerHeight - adjust */
-        let top = percentage * height - 8
-
-        const spacingPercentage = (indexHeightSum[spacingCount] - 200) / maxScrollTop
-        const spacingTop = spacingPercentage * height
-
-        sumCount += data[1]
-        spacingCount += data[1]
-        let date
-        let zIndex = 2
-        if (currentYear !== parseInt(data[0], 10)) {
-          date = parseInt(data[0], 10)
-        } else {
-          date = <hr style={{ width: 8 }} />
-        }
-        currentYear = parseInt(data[0], 10)
-        if (!index) { // first date
-          top = 8
-          spacingCount = 0
-        } else if (index === month.size - 1) { // last date
-          top += 20
-          if (top > height - 26) top = height - 26
-        } else if (spacingTop > 32 && date === parseInt(data[0], 10)) { // show years with enough spacing
-          spacingCount = 0
-        } else if (date === parseInt(data[0], 10)) { // hide years without enough spacing
-          date = null
-        } else { // show bar
-          zIndex = 1
-        }
-
-        /* set range of displaying date */
-        if (top < 16 && index) date = null
-        if (top > (height - 46) && index !== month.size - 1) date = null
-        return [date, top, zIndex, percentage]
-      })
-      return timeline
-    }
 
     this.setAnimation = (component, status) => {
       if (component === 'NavigationMenu') {
@@ -471,8 +285,6 @@ class Media extends Base {
   renderContent() {
     return (<PhotoApp
       media={this.media}
-      setPhotoInfo={this.setPhotoInfo}
-      getTimeline={this.getTimeline}
       ipcRenderer={ipcRenderer}
       apis={this.ctx.props.apis}
       requestData={this.requestData}
